[BT-66] restructure auth/user APIs and clean up form types
- Moved `currentUserApi` from `auth-api.ts` to `user-api.ts`
- Updated `ENDPOINTS` in `api-config.ts` to move `currentUser` under `user`
- Removed `registerFormDataType` and `FormDataType` from `auth-api.ts`
- Replaced `FormDataType` with `RegisterFormValues` in `useRegister` and `useGoogleRegister`
- Replaced `FormDataType` with `LoginFormData` in `useLogin`
- Created reusable `getAuthHeaders` helper in `lib/firebase.ts` for Authorization headers
- Updated fetch calls to use `getAuthHeaders` instead of repeating token logic
- Removed unnecessary `console.log` from `useGoogleSignIn` hook
