+++ app/hooks/auth/useGoogleRegister.ts
@@ -1,23 +1,14 @@
 import { googleRegisterApi } from '@app/api/auth-api';
 import routeNames from '@app/routes/route-names';
+import { RegisterFormValues } from '@app/schemas';
 import { useMutation } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
-
-interface FormDataType {
-  firstName: string;
-  lastName: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-  phoneNumber: string;
-  dateOfBirth: Date;
-}
 
 export const useGoogleRegister = () => {
   const navigate = useNavigate();
 
   const { mutate, isPending, error } = useMutation({
-    mutationFn: (formData: FormDataType) => googleRegisterApi(formData),
+    mutationFn: (formData: RegisterFormValues) => googleRegisterApi(formData),
     onSuccess: () => {
       navigate(routeNames.dashboard());
     },
